[#287] separate faq item from faq.js
